[SPARK-44227][SQL] Extract SchemaUtils from StructField.
